Address code-review findings on the ImmutableBitmap pixel path
* BitmapMemory: make ReleaseUnmanagedResources idempotent (zero Address after
  FreeHGlobal). SKBitmap.InstallPixels invokes the release proc even when it
  returns false, so the failure branch in ImmutableBitmap could dispose the same
  BitmapMemory twice -> double free / heap corruption. Idempotency also guards any
  other native-handoff consumer that double-disposes.

* Bitmap.CopyPixelsCore: compute minBufferSize in 64-bit so a very large
  stride*height can't overflow the buffer-size guard and let the new contiguous
  blit (or the per-row loop) run past the buffers.

* ImmutableBitmap: dispose the SKBitmap if SKImage.FromBitmap returns null so the
  backing memory is freed promptly via the release proc instead of waiting for the
  finalizer.

+++ src/Avalonia.Base/Media/Imaging/Bitmap.cs
@@ -213,7 +213,9 @@ namespace Avalonia.Media.Imaging
             if (stride < minStride)
                 throw new ArgumentOutOfRangeException(nameof(stride));
 
-            var minBufferSize = stride * sourceRect.Height;
+            // 64-bit to avoid overflowing the guard for very large strides/heights, which would
+            // otherwise let an oversized contiguous blit/loop run past the buffers.
+            var minBufferSize = (long)stride * sourceRect.Height;
             if (minBufferSize > bufferSize)
                 throw new ArgumentOutOfRangeException(nameof(bufferSize));
